+++ bertopic/plotting/_distribution.py
@@ -35,6 +35,9 @@ def visualize_distribution(topic_model,
     <iframe src="../../tutorial/visualization/probabilities.html"
     style="width:1000px; height: 500px; border: 0px;""></iframe>
     """
+    if len(probabilities.shape) != 2:
+        raise ValueError("This visualization cannot be used if you have set `calculate_probabilities` to False "
+                         "as it uses the topic probabilities of all topics. ")
     if len(probabilities[probabilities > min_probability]) == 0:
         raise ValueError("There are no values where `min_probability` is higher than the "
                          "probabilities that were supplied. Lower `min_probability` to prevent this error.")